+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - An endpoint to get a list of existing app names
+- Health Check endpoint
 - Application filter in Web UI
 - Custom timeframe selection with DatePicker in Web UI
 ### Fixed